+++ prisma/schema.prisma
@@ -32,6 +32,7 @@ model VisitedPark {
 }
 model User {
   id            String       @default(cuid()) @id
+  userId        String      @unique
   name          String?
   email         String?   @unique
   createdAt     DateTime  @default(now()) @map(name: "created_at")
