+++ src/app/create-profile/page.tsx
@@ -1,0 +1,9 @@
+
+
+export default function CreateProfile() {
+    return (
+        <div>
+        Create Profile
+        </div>
+    )
+}
