+++ src/app/layout.tsx
@@ -14,8 +14,6 @@ function RootLayout({
   children: React.ReactNode
 }) {
 
-  
-
   return (
     <ClerkProvider>
       <html lang="en">
